docs: update stepper
