+++ docs/demo/stepper/section-desc.jsx
@@ -2,9 +2,9 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Stepper from '../../../components/stepper'
 import Icon from '../../../components/icon'
-const prefix = 'stepper-vertical'
+const prefix = 'stepper-desc'
+const desc = '步骤名称不足以表达明确用意，用辅助信息进一步说明'
 const rightOptions = ['基础', '图标']
-const desc = '步骤与内容通过左右结构可以有效利用页面空间'
 const code = [
   {
     code: `import React from 'react'
@@ -16,31 +16,32 @@ class Demo extends React.Component {
       {
         title: '账号信息',
         content: '请输入账号信息',
+        icon: <Icon name='user' />
       },
       {
         title: '邮箱激活',
         content: '请输入邮箱',
+        icon: <Icon name='time' />
       },
       {
         title: '信息登记',
         content: '请输入个人信息',
+        icon: <Icon name='list' />
       },
     ]
 
     return (
-      <div
-        style={{height: '500px', width: '130px'}}
-      >
+      <div>
         <Stepper
           data={list}
           current={1}
-          placement='vertical'
+          itemLayout='vertical'
         />
       </div>
     )
   }
 }`,
-    opt: ['基础']
+    opt: ['图标']
   },
   {
     code: `import React from 'react'
@@ -52,44 +53,36 @@ class Demo extends React.Component {
           {
             title: '账号信息',
             content: '请输入账号信息',
-            icon: <Icon name='user' />,
           },
           {
             title: '邮箱激活',
             content: '请输入邮箱',
-            icon: <Icon name='time' />,
           },
           {
             title: '信息登记',
             content: '请输入个人信息',
-            icon: <Icon name='list' />,
           },
         ]
     
         return (
-          <div
-            style={{ height: 500 }}
-          >
-            <Stepper
-              data={list}
-              current={1}
-              placement='vertical'
-            />
-          </div>
+          <Stepper
+            data={list}
+            current={2}
+            itemLayout='vertical'
+          />
         )
       }
     }`,
-    opt: ['图标']
+    opt: ['基础']
   }
 ]
-
-const DemoVertical = () => (
+const DemoUpDownIcon = () => (
   <DocViewer
     code={code}
     scope={{ Stepper, Icon }}
+    rightOptions={rightOptions}
     prefix={prefix}
     desc={desc}
-    rightOptions={rightOptions}
   />
 )
-export default DemoVertical
+export default DemoUpDownIcon
