+++ docs/demo/stepper/section-icon.jsx
@@ -4,7 +4,10 @@ import Stepper from '../../../components/stepper'
 import Icon from '../../../components/icon'
 const desc = '每个步骤可用图标明确表示含义，带来丰富生动的效果'
 const prefix = 'stepper-icon'
-const code = `import React from 'react'
+const rightOptions = ['上下结构', '左右结构']
+const code = [
+  {
+    code: `import React from 'react'
 import Icon from '@hi-ui/hiui/es/icon'
 import Stepper from '@hi-ui/hiui/es/stepper'\n
 class Demo extends React.Component {
@@ -31,8 +34,49 @@ class Demo extends React.Component {
       />
     )
   }
-}`
+}`,
+    opt: ['左右结构']
+  },
+  {
+    code: `import React from 'react'
+import Icon from '@hi-ui/hiui/es/icon'
+import Stepper from '@hi-ui/hiui/es/stepper'\n
+class Demo extends React.Component {
+  render() {
+    const list = [
+      {
+        title: '账号信息',
+        icon: <Icon name='user' />,
+      },
+      {
+        title: '邮箱激活',
+        icon: <Icon name='time' />
+      },
+      {
+        title: '信息登记',
+        icon: <Icon name='list' />
+      },
+    ]
+
+    return (
+      <Stepper
+        itemLayout="vertical"
+        data={list}
+        current={1}
+      />
+    )
+  }
+}`,
+    opt: ['上下结构']
+  }
+]
 const DemoIcon = () => (
-  <DocViewer code={code} scope={{ Stepper, Icon }} prefix={prefix} desc={desc} />
+  <DocViewer
+    code={code}
+    scope={{ Stepper, Icon }}
+    rightOptions={rightOptions}
+    prefix={prefix}
+    desc={desc}
+  />
 )
 export default DemoIcon
