+++ docs/zh-CN/components/stepper.mdx
@@ -1,11 +1,12 @@
-# Stepper 步骤
+# Stepper 步骤条
 
 用来标识和组织完成一个任务时的信息
 
 ## 何时使用
 
-- 当完成的目标需要分布来执行，且带有明确的先后次序
-- 一个任务的每一步都需要较为复杂的信息和执行操作
+当完成的目标需要分布来执行，且带有明确的先后次序
+
+一个任务的每一步都需要较为复杂的信息和执行操作
 
 ## 基础用法
 
@@ -19,29 +20,17 @@ import DemoIcon from '../../demo/stepper/section-icon.jsx'
 
 <DemoIcon />
 
-## 上下结构
+## 辅助信息
 
-import DemoUpDown from '../../demo/stepper/section-up-down.jsx'
+import DemoDesc from '../../demo/stepper/section-desc.jsx'
 
-<DemoUpDown />
+<DemoDesc />
 
-## 上下结构的图标用法
-
-import DemoUpDownIcon from '../../demo/stepper/section-up-down-icon.jsx'
-
-<DemoUpDownIcon />
-
-## 竖直方向
+## 垂直用法
 
 import DemoVertical from '../../demo/stepper/section-vertical.jsx'
 
 <DemoVertical />
-
-## 竖直方向图标用法
-
-import DemoVerticalIcon from '../../demo/stepper/section-vertical-icon.jsx'
-
-<DemoVerticalIcon />
 
 ## Props
